move kv cache to modeling_all_models
